Implements an async web crawler
Since the input is going to come from another
concern, i.e. a cli app, the domain is isolated
on its own module.

Tests for now only target the `UriMapper` which
is responsible to transform a relative uri to
absolute according to certain criteria:
- the uri is limited ot only one subdomain.
- given ignored extensions.
- given valid schemes in uri protocol.

The `demo` module is there for now to quickly run
the crawler in memory to see how it behaves. To be
removed later when cli is adapted to the domain.